observeCafeを分割
